Move photo resize methods to helper
